second attempt I've updated the workflow to use npm install --legacy-peer-deps.
The error you saw (ERESOLVE unable to resolve dependency tree) happens because the library uses an older version of TypeScript ESLint tools that conflict with the newer ESLint version defined in its package.json.

Since we just need to build the library and not necessarily fix its linting setup, telling npm to ignore these strict peer dependency warnings with --legacy-peer-deps will allow the installation to proceed and the build to complete.

Note: I should probably fix all the warning message and outdated stuff of the lib..........